Have added and configured RabbitMq Have added InfoBoardService to send information with MQ Have added some tests Have added logging to the controllers

+++ src/main/java/ru/tsystems/javaschool/controller/manager/DriverController.java
@@ -1,5 +1,7 @@
 package ru.tsystems.javaschool.controller.manager;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -17,6 +19,7 @@ import ru.tsystems.javaschool.model.enums.Role;
 import ru.tsystems.javaschool.service.CityService;
 import ru.tsystems.javaschool.service.DriverService;
 import ru.tsystems.javaschool.service.UserService;
+import ru.tsystems.javaschool.validator.DriverValidator;
 
 import javax.validation.Valid;
 import java.time.LocalDate;
@@ -28,8 +31,17 @@ import java.util.List;
 @Controller
 public class DriverController {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(DriverController.class);
+
     private static final String DRIVER_LIST_VIEW_PATH = "redirect:/manager/listDrivers";
     private static final String ADD_DRIVER_VIEW_PATH = "newdriver";
+
+    private DriverValidator driverValidator;
+
+    @Autowired
+    public void setDriverValidator(DriverValidator driverValidator) {
+        this.driverValidator = driverValidator;
+    }
 
     private DriverService driverService;
 
@@ -52,18 +64,23 @@ public class DriverController {
         this.cityService = cityService;
     }
 
-    @RequestMapping(path = "manager/listDrivers")
+    @GetMapping(path = "manager/listDrivers")
     public String listOfDrivers(Model model) throws TruckingServiceException {
         model.addAttribute("drivers", driverService.findAllDrivers());
-        model.addAttribute("user", getPrincipal());
+        LOGGER.info("Manager {} looks on the drivers list", getPrincipal());
         return "alldrivers";
     }
 
 
     @GetMapping(path = { "manager/delete-{id}-driver" })
     public String deleteDriver(@PathVariable Integer id) throws TruckingServiceException {
+        if(driverService.findDriverById(id).getOrder()!=null){
+            LOGGER.info("Manager {} has tried to delete driver with id = {} but driver doesn't exist", getPrincipal(), id);
+            return DRIVER_LIST_VIEW_PATH;
+        }
         driverService.deleteDriver(id);
         userService.delete(id);
+        LOGGER.info("Manager {} has deleted driver with id = {}", getPrincipal(), id);
         return DRIVER_LIST_VIEW_PATH;
     }
 
@@ -71,7 +88,6 @@ public class DriverController {
     public String newDriver(ModelMap model) {
         model.addAttribute("driver", new Driver());
         model.addAttribute("edit", false);
-        model.addAttribute("user", getPrincipal());
         return ADD_DRIVER_VIEW_PATH;
     }
 
@@ -79,19 +95,25 @@ public class DriverController {
     public String saveDriver(@ModelAttribute Driver driver, BindingResult result,
                             ModelMap model) throws TruckingServiceException {
 
+        driverValidator.validate(driver, result);
+
         if (result.hasErrors()) {
+            LOGGER.info("Manager {} has tried to create a driver but entered incorrect data", getPrincipal());
             return ADD_DRIVER_VIEW_PATH;
         }
         User user = new User();
-        user.setLogin(driverService.generateDriverLogin(driver));
-        user.setPassword(driverService.generateDriverPassword());
+        String login = driverService.generateDriverLogin(driver);
+        String password = driverService.generateDriverPassword();
+        user.setLogin(login);
+        user.setPassword(password);
         user.setRole(Role.DRIVER);
         userService.save(user);
         driver.setId(user.getId());
         driver.setWorkedThisMonth(0);
         driver.setStatus(DriverStatus.REST);
         driverService.saveDriver(driver);
-
+        //driverService.sendSuccessRegistrationEmail(driver.getEmail(), login, password);
+        LOGGER.info("Manager {} has created the driver with id = {}", getPrincipal(), driver.getId());
         model.addAttribute("success", "Driver " + driver.getName()
                 + " " + driver.getSurname() + " added successfully");
         return DRIVER_LIST_VIEW_PATH;
@@ -100,10 +122,12 @@ public class DriverController {
 
     @GetMapping(value = { "manager/edit-{id}-driver" })
     public String editDriver(@PathVariable Integer id, ModelMap model) throws TruckingServiceException {
+        if(driverService.findDriverById(id).getOrder()!=null){
+            return DRIVER_LIST_VIEW_PATH;
+        }
         Driver driver = driverService.findDriverById(id);
         model.addAttribute("driver", driver);
         model.addAttribute("edit", true);
-        model.addAttribute("user", getPrincipal());
         return ADD_DRIVER_VIEW_PATH;
     }
 
@@ -111,8 +135,11 @@ public class DriverController {
     public String editDriver(@Valid Driver driver, BindingResult result,
                              ModelMap model, @PathVariable Integer id) throws TruckingServiceException {
 
+        driverValidator.validate(driver, result);
+
         if (result.hasErrors()) {
             model.addAttribute("edit", true);
+            LOGGER.info("Manager {} has tried to edit the driver but entered incorrect data", getPrincipal());
             return ADD_DRIVER_VIEW_PATH;
         }
         Driver entityDriver = driverService.findDriverById(id);
@@ -120,9 +147,11 @@ public class DriverController {
         driver.setWorkedThisMonth(entityDriver.getWorkedThisMonth());
         driverService.updateDriver(driver);
 
+        LOGGER.info("Manager {} has edited the driver with id = {}", getPrincipal(), driver.getId());
         return DRIVER_LIST_VIEW_PATH;
     }
 
+    @ModelAttribute("user")
     private String getPrincipal(){
         String userName = null;
         Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
@@ -148,7 +177,6 @@ public class DriverController {
     @ModelAttribute("date")
     public LocalDate currentDate(){
         LocalDateTime localDate = LocalDateTime.ofInstant(new Date(System.currentTimeMillis()).toInstant(), ZoneId.systemDefault());
-        LocalDate toLocalDate = localDate.toLocalDate();
-        return toLocalDate;
+        return localDate.toLocalDate();
     }
 }
